requirements.txt
# Core FastAPI framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0

# OpenAI/OpenRouter integration
openai==1.12.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
email-reply-parser==0.5.12
python-multipart==0.0.9

# HTTP client for document downloads
requests==2.31.0
httpx==0.27.0

# Data processing and ML
numpy==1.26.0
scikit-learn==1.4.0

# Environment configuration
python-dotenv==1.0.0
